% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggobsbeta3D}
\alias{ggobsbeta3D}
\title{ggplot for beta diversity}
\usage{
ggobsbeta3D(output, type = c("B", "D"), transp = 0.4)
}
\arguments{
\item{output}{the output from obsbeta3D}

\item{type}{selection of plot type : \code{type = 'B'} for plotting the gamma, alpha, and beta diversity ; \code{type = 'D'} for plotting 4 turnover dissimilarities.}

\item{transp}{a value between 0 and 1 controlling transparency. \code{transp = 0} is completely transparent, default is 0.4.}
}
\value{
a figure for Beta diversity or dissimilarity diversity.
}
\description{
\code{ggobsbeta3D}: ggplot for observed beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
data(beetle_abu)
output1 = obsbeta3D(data = beetle_abu, diversity = 'TD', datatype = 'abundance', nboot = 20)

ggobsbeta3D(output1, type = 'B')
ggobsbeta3D(output1, type = 'D')


## Taxonomic diversity for incidence data
data(beetle_inc)
output2 = obsbeta3D(data = beetle_inc, diversity = 'TD', datatype = 'incidence_raw', nboot = 20, conf = 0.95)

ggobsbeta3D(output2, type = 'B')
ggobsbeta3D(output2, type = 'D')


## Phylogenetic Hill numbers for abundance data
data(beetle_abu)
data(beetle_tree)
output3 = obsbeta3D(data = beetle_abu, diversity = 'PD', datatype = 'abundance', nboot = 20, conf = 0.95, PDtree = beetle_tree, PDreftime = NULL, PDtype = 'meanPD')

ggobsbeta3D(output3, type = 'B')
ggobsbeta3D(output3, type = 'D')


## Phylogenetic Hill numbers for incidence data
data(beetle_inc)
data(beetle_tree)
output4 = obsbeta3D(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', nboot = 0, conf = 0.95, PDtree = beetle_tree, PDreftime = NULL, PDtype = 'meanPD')

ggobsbeta3D(output4, type = 'B')
ggobsbeta3D(output4, type = 'D')


## Functional diversity for abundance data under single threshold
data(beetle_abu)
data(beetle_distM)
output5 = obsbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', nboot = 20, conf = 0.95, FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggobsbeta3D(output5, type = 'B')
ggobsbeta3D(output5, type = 'D')


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output6 = obsbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', nboot = 20, conf = 0.95, FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggobsbeta3D(output6, type = 'B')
ggobsbeta3D(output6, type = 'D')


## Functional diversity for abundance data with thresholds integrating from 0 to 1
data(beetle_abu)
data(beetle_distM)
output7 = obsbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', nboot = 10, conf = 0.95, FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggobsbeta3D(output7, type = 'B')
ggobsbeta3D(output7, type = 'D')


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = obsbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', nboot = 10, conf = 0.95, FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggobsbeta3D(output8, type = 'B')
ggobsbeta3D(output8, type = 'D')

}
