% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggObsbeta3D}
\alias{ggObsbeta3D}
\title{ggplot for beta diversity}
\usage{
ggObsbeta3D(output, type = "B")
}
\arguments{
\item{output}{the output from the function Obsbeta3D}

\item{type}{\code{type = 'B'} for plotting gamma, alpha, and beta diversity; \code{type = 'D'} for plotting four dissimilarity indices.}
}
\value{
a figure for gamma, alpha, and Beta diversity, or a figure for four dissimilarity indices.
}
\description{
\code{ggObsbeta3D}: ggplot for observed beta diversity with order q
}
\examples{
## Taxonomic diversity for abundance data
data(beetle_abu)
output1 = Obsbeta3D(data = beetle_abu, diversity = 'TD', datatype = 'abundance', nboot = 20)

ggObsbeta3D(output1, type = 'B')
ggObsbeta3D(output1, type = 'D')


## Phylogenetic Hill numbers for abundance data
data(beetle_abu)
data(beetle_tree)
output2 = Obsbeta3D(data = beetle_abu, diversity = 'PD', datatype = 'abundance', nboot = 20, conf = 0.95, 
                    PDtree = beetle_tree, PDreftime = NULL, PDtype = 'meanPD')

ggObsbeta3D(output2, type = 'B')
ggObsbeta3D(output2, type = 'D')


## Functional diversity for abundance data under single threshold
data(beetle_abu)
data(beetle_distM)
output3 = Obsbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', nboot = 20, conf = 0.95, 
                    FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggObsbeta3D(output3, type = 'B')
ggObsbeta3D(output3, type = 'D')


## Functional diversity for abundance data with thresholds integrating from 0 to 1
data(beetle_abu)
data(beetle_distM)
output4 = Obsbeta3D(data = beetle_abu, diversity = 'FD', datatype = 'abundance', nboot = 10, conf = 0.95, 
                    FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggObsbeta3D(output4, type = 'B')
ggObsbeta3D(output4, type = 'D')


## Taxonomic diversity for incidence data
data(beetle_inc)
output5 = Obsbeta3D(data = beetle_inc, diversity = 'TD', datatype = 'incidence_raw', nboot = 20, conf = 0.95)

ggObsbeta3D(output5, type = 'B')
ggObsbeta3D(output5, type = 'D')


## Phylogenetic Hill numbers for incidence data
data(beetle_inc)
data(beetle_tree)
output6 = Obsbeta3D(data = beetle_inc, diversity = 'PD', datatype = 'incidence_raw', nboot = 0, conf = 0.95, 
                    PDtree = beetle_tree, PDreftime = NULL)

ggObsbeta3D(output6, type = 'B')
ggObsbeta3D(output6, type = 'D')


## Functional diversity for incidence data under single threshold
data(beetle_inc)
data(beetle_distM)
output7 = Obsbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', nboot = 20, conf = 0.95, 
                    FDdistM = beetle_distM, FDtype = 'tau_value', FDtau = NULL)

ggObsbeta3D(output7, type = 'B')
ggObsbeta3D(output7, type = 'D')


## Functional diversity for incidence data with thresholds integrating from 0 to 1
data(beetle_inc)
data(beetle_distM)
output8 = Obsbeta3D(data = beetle_inc, diversity = 'FD', datatype = 'incidence_raw', nboot = 10, conf = 0.95, 
                    FDdistM = beetle_distM, FDtype = 'AUC', FDcut_number = 30)

ggObsbeta3D(output8, type = 'B')
ggObsbeta3D(output8, type = 'D')

}
